Moved the check of image in utils

+++ bookmarks_app/views/utils.py
@@ -1,6 +1,6 @@
 """Helper functions."""
 
-from os.path import basename
+from os.path import basename, isfile
 
 from flask import request
 import requests
@@ -14,25 +14,24 @@ def paginate(query):
     return query.paginate(page=request.args.get('page', 1), per_page=5)
 
 
-def url_has_img(url):
-    """Look if url has an image in the form of og:image in it's content."""
+def get_url_thumbnail(url):
+    """Save url's image, if does not exist already locally."""
     # TODO optimization: get chunks of data until find the og:image
     # same to the script for suggesting the title.
     response = requests.get(url)
     if response.status_code == 200:
         soup = BeautifulSoup(response.content, 'html.parser')
-        return soup.find('meta', {'property': 'og:image'})['content']
-    return None
-        
-
-def get_url_thumbnail(url):
-    """Save url's image."""
-    img_response = requests.get(url, stream=True)
-    if img_response.status_code == 200:
-        img_name = basename(url)
+        img_link =  soup.find('meta', {'property': 'og:image'})['content']
+        img_name = basename(img_link)
         destination = app.static_folder + '/img/' + img_name
-        with open(destination, 'wb') as fob:
-            for chunk in img_response:
-                fob.write(chunk)
+        if not isfile(destination):
+            img_response = requests.get(url, stream=True)
+            if img_response.status_code == 200:
+                with open(destination, 'wb') as fob:
+                    for chunk in img_response:
+                        fob.write(chunk)
+            else:
+                # TODO if not accessible i should re-try to download
+                return None
         return img_name
     return None
